Ajout de HistoChart dans la vue SeriesEvolution + introduction des SUMMARY dans BalanceStat

+++ picsou/picsou/src/main/java/org/designup/picsou/gui/series/view/SeriesWrapperUpdater.java
@@ -51,7 +51,9 @@ public class SeriesWrapperUpdater implements ChangeSetListener {
       public void visitDeletion(Key key, FieldValues previousValues) throws Exception {
         Integer seriesId = key.get(Series.ID);
         Glob wrapper = SeriesWrapper.find(localRepository, SeriesWrapperType.SERIES, seriesId);
-        localRepository.delete(wrapper.getKey());
+        if (wrapper != null) {
+          localRepository.delete(wrapper.getKey());
+        }
       }
     });
   }
